Fix tests
These tests currently prove that the library is modifying the view in a predictable manner.
However, they are not quite ideal because we're not currently able to an image snapshot working reliably. Instead we are using a recursive text description of the View. This works because it proves that the View is exactly the same each time. But it doesn't make it easy to visually see how the library is changing the look of the view.

Problem: the `image` snapshot strategy ( Snapshotting<Value, Format>) is not behaving the same as the simulator. In particular, it appears to ignore `View.preferredColorScheme()`